HZN-1510: Remove VertexHopGraphProvider and use CollapsibleGraph instead
