+++ features/topology-map/org.opennms.features.topology.api/src/main/java/org/opennms/features/topology/api/topo/SimpleSearchProvider.java
@@ -34,7 +34,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.opennms.features.topology.api.GraphContainer;
-import org.opennms.features.topology.api.support.VertexHopGraphProvider.DefaultVertexHopCriteria;
+import org.opennms.features.topology.api.support.hops.DefaultVertexHopCriteria;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
