+++ features/topology-map/org.opennms.features.topology.app/src/main/java/org/opennms/features/topology/app/internal/IpLikeSearchProvider.java
@@ -42,7 +42,7 @@ import org.opennms.core.criteria.CriteriaBuilder;
 import org.opennms.features.topology.api.GraphContainer;
 import org.opennms.features.topology.api.OperationContext;
 import org.opennms.features.topology.api.support.HistoryAwareSearchProvider;
-import org.opennms.features.topology.api.support.VertexHopGraphProvider;
+import org.opennms.features.topology.api.support.hops.CriteriaUtils;
 import org.opennms.features.topology.api.topo.AbstractSearchProvider;
 import org.opennms.features.topology.api.topo.CollapsibleCriteria;
 import org.opennms.features.topology.api.topo.SearchQuery;
@@ -316,7 +316,7 @@ public class IpLikeSearchProvider extends AbstractSearchProvider implements Hist
 	}
 
     private static CollapsibleCriteria getMatchingCriteriaById(GraphContainer graphContainer, String id) {
-        CollapsibleCriteria[] criteria = VertexHopGraphProvider.getCollapsibleCriteriaForContainer(graphContainer);
+        CollapsibleCriteria[] criteria = CriteriaUtils.getCollapsibleCriteriaForContainer(graphContainer);
         for (CollapsibleCriteria criterion : criteria) {
             if (criterion.getId().equals(id)) {
                 return criterion;
