+++ features/topology-map/org.opennms.features.topology.app/src/main/java/org/opennms/features/topology/app/internal/operations/RemoveFocusOtherVerticesOperation.java
@@ -34,9 +34,9 @@ import java.util.Set;
 import org.opennms.features.topology.api.GraphContainer;
 import org.opennms.features.topology.api.Operation;
 import org.opennms.features.topology.api.OperationContext;
-import org.opennms.features.topology.api.support.VertexHopGraphProvider;
-import org.opennms.features.topology.api.support.VertexHopGraphProvider.VertexHopCriteria;
-import org.opennms.features.topology.api.support.VertexHopGraphProvider.WrappedVertexHopCriteria;
+import org.opennms.features.topology.api.support.hops.CriteriaUtils;
+import org.opennms.features.topology.api.support.hops.VertexHopCriteria;
+import org.opennms.features.topology.api.support.hops.WrappedVertexHopCriteria;
 import org.opennms.features.topology.api.topo.Criteria;
 import org.opennms.features.topology.api.topo.VertexRef;
 
@@ -81,7 +81,7 @@ public class RemoveFocusOtherVerticesOperation implements Operation {
             return false;
         }
         final GraphContainer graphContainer = operationContext.getGraphContainer();
-        final WrappedVertexHopCriteria wrappedVertexHopCriteria = VertexHopGraphProvider.getWrappedVertexHopCriteria(graphContainer);
+        final WrappedVertexHopCriteria wrappedVertexHopCriteria = CriteriaUtils.getWrappedVertexHopCriteria(graphContainer);
         if (wrappedVertexHopCriteria.isEmpty()) {
             return false;
         }
