+++ features/topology-map/org.opennms.features.topology.app/src/main/java/org/opennms/features/topology/app/internal/service/DefaultTopologyService.java
@@ -48,7 +48,7 @@ import org.opennms.features.topology.api.topo.GraphProvider;
 import org.opennms.features.topology.api.topo.MetaTopologyProvider;
 import org.opennms.features.topology.api.topo.StatusProvider;
 import org.opennms.features.topology.api.topo.Vertex;
-import org.opennms.features.topology.api.topo.blablabla.XXXGraph;
+import org.opennms.features.topology.api.topo.blablabla.CollapsibleGraph;
 import org.opennms.features.topology.app.internal.jung.D3TopoLayoutAlgorithm;
 import org.opennms.features.topology.app.internal.operations.LayoutOperation;
 import org.opennms.netmgt.enlinkd.persistence.api.TopologyEntityCache;
@@ -152,20 +152,18 @@ public class DefaultTopologyService implements TopologyService {
             semanticZoomLevel = 0;
         }
         final GraphProvider graphProvider = getGraphProvider(metaTopologyId, namespace);
-        final XXXGraph backendGraph = graphProvider.getCurrentGraph();
+        final CollapsibleGraph collapsibleGraph = new CollapsibleGraph(graphProvider.getCurrentGraph());
 
         // Determine visible vertices and edges
         final List<Vertex> displayVertices = new ArrayList<>();
-        for (Vertex v : backendGraph.getVertices(criteria)) {
-            int vzl = backendGraph.getSemanticZoomLevel(v);
+        // TODO MVR have collapsibleGraph.getVertices() consider the szl already
+        for (Vertex v : collapsibleGraph.getVertices(criteria)) {
+            int vzl = collapsibleGraph.getSemanticZoomLevel(v);
             if (vzl <= semanticZoomLevel) {
                 displayVertices.add(v);
             }
-//            if (vzl == semanticZoomLevel || (vzl < semanticZoomLevel && !graphProvider.hasChildren(v))) {
-//                displayVertices.add(v);
-//            }
         }
-        final Collection<Edge> displayEdges = backendGraph.getEdges(criteria);
+        final Collection<Edge> displayEdges = collapsibleGraph.getEdges(criteria);
 
         // Create graph object
         final DefaultGraph uiGraph = new DefaultGraph(displayVertices, displayEdges);
@@ -174,10 +172,10 @@ public class DefaultTopologyService implements TopologyService {
         final StatusProvider vertexStatusProvider = serviceLocator != null ? findVertexStatusProvider(graphProvider) : null;
         final EdgeStatusProvider edgeStatusProvider = serviceLocator != null ? findEdgeStatusProvider(graphProvider) : null;
         if (vertexStatusProvider != null) {
-            uiGraph.setVertexStatus(vertexStatusProvider.getStatusForVertices(backendGraph, new ArrayList<>(displayVertices), criteria));
+            uiGraph.setVertexStatus(vertexStatusProvider.getStatusForVertices(collapsibleGraph, new ArrayList<>(displayVertices), criteria));
         }
         if(edgeStatusProvider != null) {
-            uiGraph.setEdgeStatus(edgeStatusProvider.getStatusForEdges(backendGraph, new ArrayList<>(uiGraph.getDisplayEdges()), criteria));
+            uiGraph.setEdgeStatus(edgeStatusProvider.getStatusForEdges(collapsibleGraph, new ArrayList<>(uiGraph.getDisplayEdges()), criteria));
         }
         return uiGraph;
     }
