+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.application/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -21,14 +21,9 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0 https://osgi.org/
 		<argument ref="applicationDao"/>
 		<property name="topologyProviderInfo" ref="applicationTopologyProviderInfo"/>
 	</bean>
-	<!-- Make it hop aware -->
-	<bean id="applicationTopologyHopProvider" class="org.opennms.features.topology.api.support.VertexHopGraphProvider">
-		<argument ref="applicationTopologyProvider" />
-	</bean>
-
     <!-- Meta topology provider -->
     <bean id="applicationMetaTopologyProvider" class="org.opennms.features.topology.api.topo.SimpleMetaTopologyProvider">
-        <argument ref="applicationTopologyHopProvider" />
+        <argument ref="applicationTopologyProvider" />
     </bean>
 
     <service interface="org.opennms.features.topology.api.topo.MetaTopologyProvider" ref="applicationMetaTopologyProvider">
