+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.pathoutage/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -10,11 +10,8 @@
         <argument ref="pathOutageStatusProvider"></argument>
         <property name="topologyProviderInfo" ref="pathoutageTopologyProviderInfo" />
     </bean>
-    <bean id="pathOutageHopProvider" class="org.opennms.features.topology.api.support.VertexHopGraphProvider">
-        <argument ref="pathOutageProvider"/>
-    </bean>
     <bean id="pathOutageMetaTopologyProvider" class="org.opennms.features.topology.api.topo.SimpleMetaTopologyProvider">
-        <argument ref="pathOutageHopProvider"/>
+        <argument ref="pathOutageProvider"/>
     </bean>
 
     <service ref="pathOutageMetaTopologyProvider" interface="org.opennms.features.topology.api.topo.MetaTopologyProvider">
