+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.pathoutage/src/test/java/org/opennms/features/topology/plugins/topo/pathoutage/PathOutageProviderTest.java
@@ -44,7 +44,7 @@ import org.mockito.Matchers;
 import org.mockito.Mockito;
 import org.opennms.core.test.OpenNMSJUnit4ClassRunner;
 import org.opennms.features.topology.api.support.FocusStrategy;
-import org.opennms.features.topology.api.support.VertexHopGraphProvider;
+import org.opennms.features.topology.api.support.hops.VertexHopCriteria;
 import org.opennms.features.topology.api.topo.Criteria;
 import org.opennms.features.topology.api.topo.Vertex;
 import org.opennms.features.topology.api.topo.blablabla.XXXGraph;
@@ -98,11 +98,11 @@ public class PathOutageProviderTest {
 	public void testHierarchy_GraphModification() {
 		this.pathOutageProvider = new PathOutageProvider(this.nodeDao, this.pathOutageStatusProvider);
 		this.pathOutageProvider.refresh();
-		List<VertexHopGraphProvider.VertexHopCriteria> criteria_original = FocusStrategy.ALL.getFocusCriteria(this.pathOutageProvider.getCurrentGraph());
+		List<VertexHopCriteria> criteria_original = FocusStrategy.ALL.getFocusCriteria(this.pathOutageProvider.getCurrentGraph());
 		this.checkProvider(criteria_original);
 		this.updateNodes();
 		this.pathOutageProvider.refresh();
-		List<VertexHopGraphProvider.VertexHopCriteria> criteria_new = FocusStrategy.ALL.getFocusCriteria(this.pathOutageProvider.getCurrentGraph());
+		List<VertexHopCriteria> criteria_new = FocusStrategy.ALL.getFocusCriteria(this.pathOutageProvider.getCurrentGraph());
 		assertThat(criteria_original, not(criteria_new));
 		this.checkProvider(criteria_new);
 	}
@@ -137,7 +137,7 @@ public class PathOutageProviderTest {
 	 * In this method the vertices from the {@link PathOutageProvider} are compared to the local vertices data
 	 * @param criteria
 	 */
-	private void checkProvider(List<VertexHopGraphProvider.VertexHopCriteria> criteria) {
+	private void checkProvider(List<VertexHopCriteria> criteria) {
 		List<Vertex> vertices = this.pathOutageProvider.getCurrentGraph().getVertices(Lists.newArrayList(criteria).toArray(new Criteria[criteria.size()]));
 		assertEquals(vertices.size(), this.nodesMap.size());
 		for (Vertex vertex : vertices) {
